requirements.txt
rembg
pillow
pyinstaller